Extract gallery detection logic into ReaderGalleryDetector
Move JS building and result parsing out of the already-large
ReaderPostDetailFragment into a focused, reusable class.
